courier_system/requirements.txt
Flask==2.2.5
flask-mysqldb==1.0.1
mysql-connector-python==8.1.0
Werkzeug==2.2.3
python-dotenv==0.19.0